requirements.txt
wheel>=0.37.1
absl-py>=0.12.0
chex>=0.0.7
dill>=0.3.3
dm-haiku>=0.0.8
# jaxline
git+https://github.com/deepmind/jaxline@f1503f6a06d46aa9eb2eab8eed6130895148ffa2
optax>=0.0.8
ml_collections
numpy>=1.16.4
numba>=0.56.3
tensorflow>=2.5.0
tensorflow-datasets>=4.3.0
dm-tree>=0.1.6
ogb>=1.3.1
# graph_nets
git+https://github.com/deepmind/graph_nets.git@64771dff0d74ca8e77b1f1dcd5a7d26634356d61
rdkit-pypi>=2021.3.2.3
jaxlib>=0.3.22+cuda11
# jraph
git+https://github.com/deepmind/jraph.git@51f5990104f7374492f8f3ea1cbc47feb411c69c
graphviz>=0.20.1
# python_graphs
git+https://github.com/google-research/python-graphs.git@0201fb090b9224dfb1a9a05ce836e9ef6da8ccc9
tbb>=2021.7.0
hydra-core>=1.3.2
